commented that FindLevelStructures is broken
The issue is that A will not necessarily be [0,-1,1,0] or [1,1,-1,0] but rather a conjugate of that. We can avoid this issue by choosing our characteristic 0 basepoint to always be j=1728 (resp. j=0), but then we will have to compute the Frobenius by hand (i.e. by using the equation x^3 + A or x^3 + Ax).