HUGE changes to generatebasicbasepoints
I've done a complete overhaul of GenerateBasicBasepoints.

First off, we are now only restricted to primes that are 1 mod 12. The reason for this is that we need p to split in both x^2+1 and x^2+x+1 - otherwise we cannot compute the Frobenius matrix. (Also, we have to choose tau to be i and zeta_3, we can't just choose some other j invariant that's congruent to 0 or 1728 mod p, otherwise we will only get automorphism matrices that are *conjugate* to [0,-1,1,0] or [1,1,-1,0].)

Second off, I've figured out the embedding problem. It turns out you really do have to compute a prime in the compositum of all Hilbert class fields. Happily, with some optimizations, this goes faster than expected.